Install as a Windows Service
The installer now installs a Windows Service, called "ChromaTemp Service," that runs a thread with the 3-second polling loop. It's not started by default. You need to go into Windows Services to enable it.

ChromaTempService is now the main executable. The "ChromaTemp" subproject is now a DLL, and ChromaTempService depends on it.

Some cleanup and renaming is needed. For example, ChromaTemp's Program.cs doesn't really make sense anymore, since it's no longer a .exe. ChromaTemp and ChromaTempService both depend on Colore, when only one dependency should be needed.